refactor(analytics): split AnalyticsService into export + leaderboard services (#753)
Decompose the 1,801-line AnalyticsService into three focused services:

- AnalyticsExportService: CSV/XLSX export + the five platform-integration
  deps (youtube/tiktok/instagram/pinterest/twitter) that only export used.
  Removes those five injections from AnalyticsService.
- EntityLeaderboardService: org/brand leaderboards + paginated stats. The
  four near-identical leaderboard methods now share one parameterized
  loadEntitiesWithStats() scaffolding and a single sortByMetric() comparator
  (was copy-pasted four times).
- AnalyticsService: keeps the dashboard/time-window methods. Constructor
  drops to prisma/configService/logger.

Oversized methods (getViralHooks, getTimeSeriesData, getOverview,
getGrowthTrends, getTopContent) decomposed into private query/mapping helpers;
every $queryRaw body is byte-identical to before (no SQL/behavior change,
perf work stays out of scope per #626/#629/#630).

Controller re-points the export + four leaderboard routes to the new services;
module registers all three. Specs split: dashboard-only analytics.service.spec,
plus colocated analytics-export.service.spec and entity-leaderboard.service.spec;
controller spec re-pointed.

Refs #753
